requirements.txt
aiohttp==3.9.5
aiokafka==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
beanie==1.26.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
deprecation==2.1.0
df2gspread==1.0.4
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
environs==11.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==1.6.7
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
jwcrypto==1.5.6
kafka-python==2.0.2
lazy-model==0.2.0
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mongoengine==0.28.2
motor==3.5.1
multidict==6.0.5
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pluggy==1.5.0
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic-settings==2.3.4
pydantic_core==2.18.3
PyDispatcher==2.0.7
Pygments==2.18.0
pymongo==4.7.2
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==4.1.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
Scrapy==2.11.2
sentry-sdk==2.11.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
slack_sdk==3.27.2
sniffio==1.3.1
starlette==0.37.2
tldextract==5.1.2
toml==0.10.2
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uritemplate==3.0.1
urllib3==2.2.2
uvicorn==0.30.0
uvloop==0.19.0
w3lib==2.2.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
zope.interface==6.4.post2
